Move guest link types to a separate file
