+++ types/types.go
@@ -25,20 +25,6 @@ type (
 		UploadMetadata
 		Reader io.ReadSeeker
 	}
-
-	GuestLinkID             string
-	GuestLinkLabel          string
-	GuestUploadMaxFileBytes uint64
-	GuestUploadCountLimit   int
-
-	GuestLink struct {
-		ID                   GuestLinkID
-		Label                GuestLinkLabel
-		Created              time.Time
-		Expires              ExpirationTime
-		MaxFileBytes         *GuestUploadMaxFileBytes
-		UploadCountRemaining *GuestUploadCountLimit
-	}
 )
 
 // Treat a distant expiration time as sort of a sentinel value signifying a "never expire" option.
@@ -46,20 +32,4 @@ var NeverExpire = ExpirationTime(time.Date(2999, time.December, 31, 0, 0, 0, 0, 
 
 func (et ExpirationTime) String() string {
 	return (time.Time(et)).String()
-}
-
-func (gl GuestLink) CanAcceptMoreFiles() bool {
-	if gl.UploadCountRemaining == nil {
-		return true
-	}
-	r := int(*gl.UploadCountRemaining)
-	return r > 0
-}
-
-func (gl *GuestLink) DecrementUploadCount() {
-	if gl.UploadCountRemaining == nil {
-		return
-	}
-	s := GuestUploadCountLimit(int(*gl.UploadCountRemaining) - 1)
-	gl.UploadCountRemaining = &s
 }
